Update ctk_tabview.py
Adding two methods to the class CTkTabview":
"index" to return the index (position) of the current tab or the tab entered as an attribute.
"len" method to return the number of defined tabs

Modifying the "get" method: it now returns the name of the selected tab, or an empty string if no tab is selected. If an index is defined, it returns the name associated with that index.