[feat] Event Generation on Canvas
For widgets based on the _canvas, attempting to use event_generate for
custom events failed. Typical notation in tkinter would allow such
bindings but in customtkinter, operations are performed on hidden
widgets. Thus a function override is required to operate as expected.

Widgets that are not based on _canvas or where a combination of hidden
widgets were not updated.

Signed off by Trask Crane (hcrane3@gatech.edu)