[feat] Combobox Validation
Allows for validation of the combobox through the hidden entry widget.
User is able to validate provided that they *do not delete or insert
text*. If the user wishes to do this in validation, they first must
ensure that the customtkinter library does not trigger the call
repeatidly AND should add a short delay (via self.after). This is
because on when `self._entry.insert` is called within customtkinter,
it will clear the `validate` property thus causing future validation
to fail.

Signed off by Trask Crane (hcrane3@gatech.edu)