Fix focus issue in CTkEntry widget
This commit addresses a bug where the CTkEntry widget was unable to lose focus when clicked outside. The `_on_root_click` method has been updated to only set the focus to the root window if the CTkEntry widget currently has the focus. This prevents the focus from being set to the root window when clicking on another widget that should receive the focus. Exception handling has also been improved for better error resilience.